requirements.txt
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.45.1
future==0.18.3
idna==3.6
importlib-resources==6.1.1
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib
matplotlib-inline==0.1.6
numpy
packaging==23.2
pandas
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scipy
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vertica-highcharts==0.1.3
vertica-python==1.3.7
verticapy
wcwidth==0.2.12
zipp==3.17.0
lxml==4.9.4
beautifulsoup4==4.12.2
patool
geopandas